Hook strong resources generator in model
