+++ lib/strong_resources/railtie.rb
@@ -1,0 +1,11 @@
+require 'active_model/railtie'
+
+module ActiveModel
+  class Railtie < Rails::Railtie
+    generators do |app|
+      Rails::Generators.configure! app.config.generators
+
+      require_relative '../generators/model_override'
+    end
+  end
+end
